requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
pyyaml>=6.0
requests>=2.31.0
duckdb>=0.9.0
pytest>=7.4.0
dash>=2.14.0
plotly>=5.17.0
jinja2>=3.1.0
gunicorn>=21.2.0
flask-httpauth>=4.8.0
vectorbt>=0.27.0
tabulate>=0.9.0
prefect>=2.14.0
scipy>=1.11.0
scikit-learn>=1.3.0
ruamel.yaml>=0.17.0
GitPython>=3.1.0